backend/src/requirements.txt
beautifulsoup4==4.10.0
boto3==1.20.16
botocore==1.23.16
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
google==3.0.0
google-api-core==2.2.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
integv==1.3.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
oauthlib==3.1.1
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.27
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
